fix: applied pr suggestions
