+++ src/functions/mock-credentials.ts
@@ -10,11 +10,12 @@ import {
   buildCertPath,
   buildJwksPath,
   ensureDir,
-  EXPIRY_LEEWAY_MS,
+  CLOCK_SKEW_TOLERANCE_MS,
   hasTrustChainExpired,
   hasX509CertificateExpired,
   loadCertificate,
   loadJwks,
+  VALIDITY_MS,
 } from "@/logic";
 import {
   Config,
@@ -53,7 +54,7 @@ export async function createMockMdlMdoc(
     subject,
   );
 
-  const expiration = new Date(Date.now() + 24 * 60 * 60 * 1000 * 365);
+  const expiration = new Date(Date.now() + VALIDITY_MS);
   let mockedMdoc: Credential;
   switch (version) {
     case ItWalletSpecsVersion.V1_0:
@@ -112,7 +113,7 @@ export async function createMockSdJwt(
     "CN=test_issuer",
   );
 
-  const expiration = new Date(Date.now() + 24 * 60 * 60 * 1000 * 365);
+  const expiration = new Date(Date.now() + VALIDITY_MS);
   let mockedSdjwt: Credential;
   switch (version) {
     case ItWalletSpecsVersion.V1_0:
@@ -228,11 +229,13 @@ export function isCredentialMdocExpired(
   const now = Date.now();
   const isCredentialExpired =
     path !== undefined &&
-    getCredentialMdocExpiration(document, path).getTime() < now;
+    getCredentialMdocExpiration(document, path).getTime() <
+     now - CLOCK_SKEW_TOLERANCE_MS;
 
   const exp =
     document.issuerSigned.issuerAuth.decodedPayload.validityInfo.validUntil.getTime();
-  const isMDocExpired = checks.mdoc && exp < now;
+  const isMDocExpired = checks.mdoc && exp <
+   now - CLOCK_SKEW_TOLERANCE_MS;
 
   const isCertExpired =
     checks.cert &&
@@ -283,14 +286,14 @@ export function isCredentialSdJwtExpired(
   const isCredentialExpired =
     claimName !== undefined &&
     getCredentialSdJwtExpiration(parsed, claimName).getTime() <
-      now - EXPIRY_LEEWAY_MS;
+      now - CLOCK_SKEW_TOLERANCE_MS;
 
   const exp = parsed.jwt.payload?.exp;
   const isJwtExpired =
     checks.jwt &&
     exp !== undefined &&
     typeof exp === "number" &&
-    exp * 1000 < now - EXPIRY_LEEWAY_MS;
+    exp * 1000 < now - CLOCK_SKEW_TOLERANCE_MS;
 
   const jwt_trust_chain = zTrustChain.safeParse(parsed.jwt.header?.trust_chain);
   const isTrustChainExpired =
